refactor: collect zeroApprove files into module
