+++ apps/cowswap-frontend/src/common/containers/TradeApprove/TradeApproveButton.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useState } from 'react'
 
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
-import { useShouldZeroApprove } from 'common/hooks/useShouldZeroApprove'
-import { useZeroApprove } from 'common/hooks/useZeroApprove'
+import { useShouldZeroApprove, useZeroApprove } from 'modules/zeroApproval'
+
 import { ApproveButton } from 'common/pure/ApproveButton'
 import { CowModal } from 'common/pure/Modal'
 import { TransactionErrorContent } from 'common/pure/TransactionErrorContent'
