--- apps/cowswap-frontend/src/common/hooks/useShouldZeroApprove/index.ts
@@ -1,1 +1,0 @@
-export * from './useShouldZeroApprove'
