+++ apps/cowswap-frontend/src/modules/trade/containers/TradeWidget/index.tsx
@@ -21,10 +21,10 @@ import { useIsWrapOrUnwrap } from 'modules/trade/hooks/useIsWrapOrUnwrap'
 import { RecipientAddressUpdater } from 'modules/trade/updaters/RecipientAddressUpdater'
 import { TradeFormValidationUpdater } from 'modules/tradeFormValidation'
 import { TradeQuoteUpdater } from 'modules/tradeQuote'
+import { ZeroApprovalModal } from 'modules/zeroApproval'
 
 import { TradeApproveModal } from 'common/containers/TradeApprove'
 import { tradeApproveStateAtom } from 'common/containers/TradeApprove/tradeApproveStateAtom'
-import { ZeroApprovalModal } from 'common/containers/ZeroApprovalModal'
 import { useCategorizeRecentActivity } from 'common/hooks/useCategorizeRecentActivity'
 import { useIsProviderNetworkUnsupported } from 'common/hooks/useIsProviderNetworkUnsupported'
 import { useModalState } from 'common/hooks/useModalState'
