+++ apps/cowswap-frontend/src/modules/zeroApproval/containers/ZeroApprovalModal/ZeroApprovalModal.cosmos.tsx
@@ -1,0 +1,13 @@
+import { ZeroApprovalModal } from './ZeroApprovalModal'
+
+const modalState = {
+  isModalOpen: true,
+  openModal: () => {},
+  closeModal: () => {},
+}
+
+const Fixtures = {
+  default: <ZeroApprovalModal modalState={modalState} />,
+}
+
+export default Fixtures
