+++ apps/cowswap-frontend/src/modules/zeroApproval/hooks/useShouldZeroApprove/index.ts
@@ -1,0 +1,2 @@
+export { useShouldZeroApprove } from './useShouldZeroApprove'
+export { shouldZeroApprove } from './shouldZeroApprove'
